refactor: update package structure and enhance UI logic for meal suggestions
