+++ src/main/kotlin/di/UseCaseModule.kt
@@ -5,20 +5,22 @@ import logic.usecase.SuggestHighCalorieMealUseCase
 import org.koin.dsl.module
 
 val useCaseModule = module {
-    single { FilterQuickHealthyMealsUseCase(get()) }
-    single { IraqiMealsIdentifierUseCase(get()) }
-    single { SearchMealUseCase(get()) }
-
     single { EggFreeSweetsUseCase(get()) }
+    single { ExploreOtherCountriesFoodCultureUseCase(get()) }
+    single { FilterQuickHealthyMealsUseCase(get()) }
+
+    single { GuessMealGameUseCase(get()) }
+    single { IngredientGameUseCase(get()) }
+    single { IraqiMealsIdentifierUseCase(get()) }
+
+    single { MealsForLargeGroupUseCase(get()) }
+    single { PotatoLovingMealsUseCase(get()) }
+    single { SeaFoodMealUseCase(get()) }
+
+    single { SearchMealUseCase(get()) }
     single { SuggestEasyMealUseCase(get()) }
     single { SuggestHighCalorieMealUseCase(get()) }
 
     single { SuggestKetoMealUseCase(get()) }
-    single { GuessMealGameUseCase(get()) }
-    single { SeaFoodMealUseCase(get()) }
-
-    single { IngredientGameUseCase(get()) }
-
-    single { PotatoLovingMealsUseCase(get()) }
 
 }
