+++ src/main/kotlin/logic/usecase/ExploreOtherCountriesFoodCultureUseCase.kt
@@ -1,4 +1,4 @@
-package logic.use_case
+package logic.usecase
 
 import logic.MealsProvider
 import model.Meal
