+++ src/main/kotlin/presentation/ExploreOtherCountriesUI.kt
@@ -1,13 +1,23 @@
 package org.example.presentation
 
+import logic.usecase.ExploreOtherCountriesFoodCultureUseCase
+
 class ExploreOtherCountriesUI(
-//    private val useCase: ExploreOtherCountriesUseCase
+    private val useCase: ExploreOtherCountriesFoodCultureUseCase
     ) : Feature {
     override val id: Int = FEATURE_ID
     override val name: String = FEATURE_NAME
 
     override fun execute() {
-        // Implement your logic here
+        print("Enter country name to discover it's culture ")
+        val countryName = readlnOrNull()?.trim().orEmpty()
+        val meals = useCase.execute(countryName)
+        println(
+            if (meals.isEmpty()) "No meals found matching '$countryName'."
+            else "Matching Meals:\n${
+                meals.joinToString("\n") { "- ${it.name}" }
+            }"
+        )
     }
 
     companion object {
