+++ src/main/kotlin/presentation/SeafoodMealsUI.kt
@@ -9,11 +9,16 @@ class SeafoodMealsUI(
     override val name: String = FEATURE_NAME
 
     override fun execute() {
-        useCase.execute()
+        val rankedMeals = useCase.execute().take(NUMBER_OF_ITEMS)
+        println("Rank    Name    Protein")
+        rankedMeals.forEach { (rank, meal) ->
+            println("$rank      ${meal.name}     ${meal.nutrition!!.protein}")
+        }
     }
 
     companion object {
         const val FEATURE_ID = 14
         const val FEATURE_NAME = "Show seafood meals sorted by protein"
+        const val NUMBER_OF_ITEMS = 50
     }
 }
